First check in of the 2.0.2 release. Main changes: - introduced FF_ESAVER (Energy SAVER) compiler flag (experimental) (thanks to Mehdi Goli ) - ff_relax now uses nanosleep instead of usleep - first full porting to ARM processor (thanks to Mauro Mulatero) - added new interfaces for the ff_farm - added a new construct ff_pipe only for C++11 - fixed some problems with cmake compilation - added ff_send_out_to as a method of the ff_loadbalancer class - added parallel_for.hpp which implements 'OpenMP-like' parallel_for - added cleanup_workers method in the farm skeleton - removed fallback function in the farm emitter - removed many warnings when compiling with the Intel compiler - fixed get_my_id() problem for pipeline

+++ tests/test_multi_input.cpp
@@ -50,7 +50,7 @@ class Emitter1: public ff_node {
 public:
     Emitter1(int ntasks):ntasks(ntasks) {}
     void* svc(void*) {
-        for(int i=0;i<ntasks;++i)
+        for(long i=0;i<ntasks;++i)
             ff_send_out((void*)(i+10));
         return NULL;
     }
